removing po catalog
